% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3BarChart.R
\name{C3BarChart}
\alias{C3BarChart}
\title{C3BarChart}
\usage{
C3BarChart(dataset, ratio, colors, width = NULL, height = NULL)
}
\arguments{
\item{dataset}{the data to be plotted, a list or data.frame containing the different data series,
each as a vector of numeric values. Note that a vector of x values must be supplied and named 'x'.}

\item{ratio}{the ratio between the width of bars and
the width of spaces in between bars, a single numeric value.}

\item{colors}{vector of colors for the bars.}

\item{width, height}{size of figure output}
}
\description{
Creat a bar chart
}
