% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3Gauge.R
\name{C3Gauge}
\alias{C3Gauge}
\title{C3Gauge}
\usage{
C3Gauge(value, width = NULL, height = NULL)
}
\arguments{
\item{value}{a single numeric value, the value to display on the gauge}

\item{width, height}{size of figure output}
}
\description{
Creat a dashboard style gauge
}
