% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3Scatter.R
\name{C3Scatter}
\alias{C3Scatter}
\title{C3Scatter}
\usage{
C3Scatter(dataset, size = 2.5, colors, width = NULL, height = NULL)
}
\arguments{
\item{dataset}{the data to be plotted, a list or data.frame containing the different data series,
each as a vector of numeric values. Note that a vector of x values must be supplied and named 'x'.}

\item{size}{the ratio of the points to be plotted}

\item{colors}{vector of colors for the bars.}

\item{width, height}{size of figure output}
}
\description{
Creat a scatter plot
}
