% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3Pie.R
\name{C3Pie}
\alias{C3Pie}
\title{C3Pie}
\usage{
C3Pie(values, legendPosition = "bottom", width = NULL, height = NULL)
}
\arguments{
\item{values}{the data to be plotted, a numeric vector containing the data to be pied.}

\item{legendPosition}{the position of the legend, defaults to "bottom"}

\item{width, height}{size of figure output}
}
\description{
Creat a pie chart
}
